requirements.txt
python-dotenv==1.1.1
langchain_openai==0.3.32
langgraph==0.6.6
langgraph-checkpoint-postgres==2.0.23
mcp==1.13.1
fastmcp==2.12.0
langchain-mcp-adapters==0.1.9
requests==2.32.5
fastapi==0.116.1
uvicorn==0.35.0
httpx==0.28.1
langgraph-supervisor==0.0.29